pyproject.toml
[tool.poetry]
name = "mbox"
version = "0.6.0"
description = "Music Box is a discord streaming bot designed to use no commands"
authors = ["mbox-contributors"]
license = "n"

[tool.poetry.dependencies]
python = "^3.7"
discord-py-slash-command = "1.2.2"
"discord.py" = ">=1.6.0"
requests = ">=2.25.0"
youtube-search = ">=1.1.1"
ytmusicapi = ">=0.15.1"
PyNaCl = "1.4.0"
youtube_dl = "2021.4.17"
python-dotenv = "0.17.1"
discord-components = "0.5.3"
rsa = "4.7.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-xprocess = "^0.17.1"
flake8 = "^3.9.2"
black = "^21.6b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.pytest.ini_options]
